server: update run.sh

+++ server/run.sh
@@ -1,7 +1,11 @@
 if cargo watch --help &> /dev/null; then
     TRIGGER="./in/datasets_len.txt"
-    echo "0" > $TRIGGER
-    cargo watch --no-vcs-ignores -w "./src" -w "./run.sh" -w "$TRIGGER" -x "run -r"
+
+    if [ ! -f "$TRIGGER" ]; then
+        echo "0" > $TRIGGER
+    fi
+
+    cargo watch --no-vcs-ignores -w "./src" -w "$TRIGGER" -x "run -r"
 else
     cargo run -r
 fi
